refactor: extract config and CLI packages from main.go (#13)
Split cmd/bugbarn/main.go (726 lines → 369 lines) into:
- internal/config: Config struct and env/file loading
- internal/cli: user, project, and apikey subcommand handlers

Pure structural refactor with no behavior changes.
